requirements.txt
wheel
pandas
discord.py
python-dotenv
